# Some more rust rayon perf updates