--- a/rust_rayon/Cargo.toml
+++ b/rust_rayon/Cargo.toml
@@ -11,8 +11,11 @@ rayon = "1.8.0"
 rustc_data_structures = "0.0.1"
 serde =  { version = "1.0.188", features = ["derive"] }
 serde_json = "1.0.107"
+smallstr = { version = "0.3.0", features = ["serde"] }
 
 [profile.release]
 lto = true
 codegen-units = 1
 debug = "line-tables-only"
+panic = "abort"
+