cuts a few MS on my machine, give it a try.