@jinyus 2023-09-28
They are pretty much equal on my machine.

New version:
```
  Benchmark 1: ./target/release/rust_rayon
  Processing time (w/o IO): 8.293485ms
  Processing time (w/o IO): 7.605749ms
  Processing time (w/o IO): 8.264292ms
  Processing time (w/o IO): 9.399018ms
  Processing time (w/o IO): 7.947556ms
  Processing time (w/o IO): 7.607355ms
  Processing time (w/o IO): 7.735027ms
  Processing time (w/o IO): 7.245502ms
  Processing time (w/o IO): 7.720779ms
  Processing time (w/o IO): 11.39029ms
  Processing time (w/o IO): 8.001615ms
  Processing time (w/o IO): 8.237123ms
  Processing time (w/o IO): 9.120976ms
    Time (mean ± σ):      25.1 ms ±   1.0 ms    [User: 69.8 ms, System: 37.0 ms]
    Range (min … max):    23.9 ms …  27.1 ms    10 runs
```

Old version:

```Benchmark 1: ./target/release/rust_rayon
Processing time (w/o IO): 10.935896ms
Processing time (w/o IO): 10.097231ms
Processing time (w/o IO): 8.831271ms
Processing time (w/o IO): 9.922276ms
Processing time (w/o IO): 8.182506ms
Processing time (w/o IO): 8.59255ms
Processing time (w/o IO): 8.346426ms
Processing time (w/o IO): 7.574948ms
Processing time (w/o IO): 8.43757ms
Processing time (w/o IO): 8.49917ms
Processing time (w/o IO): 8.656455ms
Processing time (w/o IO): 8.069641ms
Processing time (w/o IO): 8.679852ms
  Time (mean ± σ):      25.0 ms ±   1.3 ms    [User: 74.1 ms, System: 36.7 ms]
  Range (min … max):    23.5 ms …  27.9 ms    10 runs
```

Plus the old version is easier to understand