@masmullin2000 2023-09-28
hmm.  no worries  Here is on my 5950x

new:
```
Benchmark 1: ./target/release/rust_rayon
Processing time (w/o IO): 5.596644ms
Processing time (w/o IO): 4.881927ms
Processing time (w/o IO): 4.509504ms
Processing time (w/o IO): 4.819202ms
Processing time (w/o IO): 4.342627ms
Processing time (w/o IO): 4.42138ms
Processing time (w/o IO): 4.41712ms
Processing time (w/o IO): 4.403669ms
Processing time (w/o IO): 4.448041ms
Processing time (w/o IO): 4.750723ms
Processing time (w/o IO): 4.558889ms
Processing time (w/o IO): 4.653067ms
Processing time (w/o IO): 4.627708ms
  Time (mean ± σ):      15.8 ms ±   0.2 ms    [User: 39.0 ms, System: 41.4 ms]
  Range (min … max):    15.5 ms …  16.3 ms    10 runs

Checking output
Output file: related_posts_rust_rayon.json is valid!
```

Old

```
Benchmark 1: ./target/release/rust_rayon
Processing time (w/o IO): 5.226727ms
Processing time (w/o IO): 4.597839ms
Processing time (w/o IO): 4.990479ms
Processing time (w/o IO): 4.561113ms
Processing time (w/o IO): 4.713648ms
Processing time (w/o IO): 6.99248ms
Processing time (w/o IO): 5.722777ms
Processing time (w/o IO): 5.201569ms
Processing time (w/o IO): 4.842887ms
Processing time (w/o IO): 6.52187ms
Processing time (w/o IO): 5.22804ms
Processing time (w/o IO): 5.136928ms
Processing time (w/o IO): 4.455229ms
  Time (mean ± σ):      22.3 ms ±   5.8 ms    [User: 67.2 ms, System: 38.4 ms]
  Range (min … max):    16.3 ms …  35.7 ms    10 runs

Checking output
Output file: related_posts_rust_rayon.json is valid!
```